WP-48: iOS Profile/-modul + demo/mock-karantene
Assistant/ var to domener i én mappe — 12 av 24 filer var profil-domenet,
plassert der fordi mappen alt var wiret i test-targetet. Ren strukturflytt,
ingen logikkendringer:

- Zenji/Profile/ (git mv, historikk følger med): ProfileStore, InterestProfile,
  ProfileMerge, ProfileSyncModel, ProfileSyncBackend, ProfileSyncCoordinator,
  CloudKitProfileSync, ProfileShareCodec, ProfileSharePanel, ProfileQRCode,
  ResetService + EffectiveInterests (broen blir i Profile/: den leser
  InterestProfile og trenger EntityIndex — Feed/ ville dratt begge inn i
  widget-targetet, som med vilje ikke har profil-kode).
- AssistantViewModel:391-480 (MARK: WP-19 profil-sync) → ny
  Profile/AssistantViewModel+ProfileSync.swift. De to LAGREDE feltene
  (lastImportSummary/shareImportMessage) blir i klassekroppen (en extension
  kan ikke holde stored state); profile/presentToken/profileStore får
  intern (ikke private) tilgang med begrunnelse i kommentar — det mekaniske
  minimum for fil-splitten.
- Zenji/Demo/ for LensDemoSeed + MemoryDemoSeed — strukturelt utenfor
  widget-/test-targets (de lister kilder eksplisitt); MemoryDemoSeeds
  header-påstand «test-target plukker den ikke opp» var feil (Zenji/Memory
  kompileres inn i ZenjiTests) og er rettet — nå stemmer den.
- MockInterestAssistant + MockAnswerer pakket i #if DEBUG (kompilertes inn i
  Release). MockMemoryDistiller likeså — dens endOfOsloDay brukes av den
  ekte FM-distillasjonen og er flyttet til ny MemoryFreshness (utenfor
  mocken, samme oppførsel; mocken beholder en forwarder så testene er
  uendret). Akseptkriteriet er «ingen Mock*-symboler i Release», derfor
  dekkes alle tre, ikke bare de to opplistede.
- project.yml: +1 kildelinje (Zenji/Profile) i ZenjiTests, med
  per-linje-begrunnelse i husstilen; app/ZenjiDeviceDev urørt (path: Zenji).

Bevis: 376/376 iOS-tester grønne (identisk med baseline før flytting);
npm test 373/373 urørt; Zenji + ZenjiWidgetExtension + ZenjiDeviceDev
bygger; nm på Release-binærene (app + widget-appex): 0 Mock*-symboler
(Debug-dylib som positiv kontroll: 193).
