+++ ios/Zenji/Profile/ProfileQRCode.swift
